Tutorial_1_docker/ejercicio 1/requirements.txt
slackclient==2.9.1
slackeventsapi==2.2.1